requirements.txt
pytest
python-dotenv
docstring_parser
deepeval
aisuite
matplotlib
numpy
